pyproject.toml
[project]
name = "website-query-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "google-genai>=1.12.1",
    "langchain-community>=0.3.22",
    "langchain-google-genai>=2.1.3",
    "langchain-qdrant>=0.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "termcolor>=3.0.1",
]
